+++ django_project/geohosting/src/pages/CheckoutPage/CheckoutFinish/index.tsx
@@ -35,7 +35,7 @@ const CheckoutConfiguration: React.FC = () => {
 
   useEffect(() => {
     // Check the url and redirect to correct page
-    if (salesOrderDetail) {
+    if (salesOrderDetail && salesOrderDetail.id === id) {
       checkCheckoutUrl(salesOrderDetail)
     }
   }, [salesOrderDetail]);
